the max_alpha value must be 1 not 255

+++ rxanimutil/src/main/java/com/anwesome/ui/rxanimutil/RxAnimUtil.java
@@ -85,7 +85,7 @@ public class RxAnimUtil {
             }
         });
     }
-    private static void fadeInOrOutView(final View view,int timeInMilliSeconds,final int dir,final int start,final int end) {
+    private static void fadeInOrOutView(final View view,int timeInMilliSeconds,final int dir,final float start,final float end) {
         view.setAlpha(start);
         animateView(timeInMilliSeconds, new RxAnimationListener() {
             @Override
